Get actual number of projects per language
There was a change sometime in the past that resulted in the pagination
of projects no longer being a good proxy for the number of projects in
a given language.  This patch utilizes the github language specific
search to get the number of projects in a given language.

It also corrects a bug with xpath processing.

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>